Extract Constructor into its own file.
